translations

+++ src/Domain/Abstraction/Security/IAuthenticationManager.cs
@@ -5,7 +5,7 @@ namespace Domain.Abstraction.Security
     public interface IAuthenticationManager
     {
         bool Authenticate(string token, out ClaimsPrincipal credentials);
-        void SignOut(string token, long userId);
+        void SignOut(string token, int userId);
         (IToken AccessToken, IToken RefreshToken) GenerateTokens(IUserCredentials credentials);
         void SetRefreshToken(IToken refreshToken);
         string GetRefreshToken();
